Moved hooks to separate files
